live_streaming_server/requirements.txt
# Core Flask dependencies (inherited from main backend)
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# WebSocket and real-time dependencies
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.21

# Cache
redis==5.0.1
Flask-Caching==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# Utilities
cryptography==41.0.4
Pillow==10.0.1

# Development and testing
pytest==7.4.2
pytest-flask==1.3.0

# Production server
gunicorn==21.2.0
eventlet==0.33.3 